+++ backend/db/models/spot.js
@@ -67,7 +67,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lat: {
-      type: DataTypes.DECIMAL(10, 7),
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
         min: -90,
@@ -75,7 +75,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lng: {
-      type: DataTypes.DECIMAL(10, 7),
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
         min: -180,
